refactors moment delete, styles journal entries, and adds entries to charts
